==> code/01_content_analysis.Rmd <==
---
title: "Content analysis"
author: "Jae Yeon Kim"
html_document:
  toc: True
  theme: united
  number_sections: True
---

# Setup

```{r, include = FALSE}

pacman::p_load(
  tidyverse, # for the tidyverse framework
  irr, # for calculating inter-labeler reliability scores
  vcd, # for kappa statistics
  corrplot, # for visualizing correlation coefficients
  ggpubr, # for arranging ggplots
  ggthemes, # for fancy ggplot themes
  gmodels, # quick summaries
  patchwork # for arranging ggplots
)


# devtools::install_github("jaeyk/makereproducible")
library(makereproducible)

# Import R scripts

script_list <- list.files(paste0(here::here(), "/functions"),
  pattern = "*.r|*.R",
  full.names = TRUE
)


for (i in 1:length(script_list))
{
  source(script_list[[i]])
}

# for publication-friendly theme
theme_set(theme_pubr())
```

# Importing files 


```{r include=FALSE}

# Training data for classifying articles from Asian American newspapers
training_asian <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/raw_data/training_asian.csv")) %>%
  dplyr::select(-contains("Notes")) %>%
  mutate(
    linked_progress_gran = ifelse(Promoting_collective_gains_A + Promoting_collective_gains_B >= 1, 1, 0),
    linked_hurt_gran = ifelse(Preventing_collective_losses_A + Preventing_collective_losses_B >= 1, 1, 0),
    linked_progress = ifelse(Promoting_collective_gains_A + Promoting_collective_gains_B == 2, 1, 0),
    linked_hurt = ifelse(Preventing_collective_losses_A + Preventing_collective_losses_B == 2, 1, 0)
  )

# Training data for classifying articles from African American newspapers
training_black <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/raw_data/training_black.csv")) %>%
  dplyr::select(-contains("Notes")) %>%
  mutate(
    linked_progress_gran = ifelse(Promoting_collective_gains_A + Promoting_collective_gains_B >= 1, 1, 0),
    linked_hurt_gran = ifelse(Preventing_collective_losses_A + Preventing_collective_losses_B >= 1, 1, 0),
    linked_progress = ifelse(Promoting_collective_gains_A + Promoting_collective_gains_B == 2, 1, 0),
    linked_hurt = ifelse(Preventing_collective_losses_A + Preventing_collective_losses_B == 2, 1, 0)
  )

# Dimensions
dim(training_asian)
dim(training_black)

# Rename variables
names(training_asian) == names(training_black)
```

# Descriptive statistics 

## Merging

```{r}

# Wrangling

df <- bind_rows(
  mutate(training_asian, group = "Asian Americans"),
  mutate(training_black, group = "African Americans")
)
```

### Class balance 

```{r}

ggarrange(
  df %>%
    group_by(group, linked_progress) %>%
    add_count() %>%
    ggplot(aes(x = as.factor(group), y = n, fill = as.factor(linked_progress))) +
    geom_col(position = "dodge", color = "black") +
    scale_fill_manual(name = "Label", labels = c("No", "Yes"), values = c("red", "blue")) +
    geom_text(aes(label = n), position = position_dodge(width = 0.9), vjust = -0.25) +
    labs(
      x = "Group", y = "Frequency",
      title = "Collective gain",
      fill = "Group"
    ),

  df %>%
    group_by(group, linked_hurt) %>%
    add_count() %>%
    ggplot(aes(x = as.factor(group), y = n, fill = as.factor(linked_hurt))) +
    geom_col(position = "dodge", color = "black") +
    scale_fill_manual(name = "Label", labels = c("No", "Yes"), values = c("red", "blue")) +
    geom_text(aes(label = n), position = position_dodge(width = 0.9), vjust = -0.25) +
    labs(
      x = "Group", y = "Frequency",
      title = "Collective loss",
      fill = "Group"
    ),
  common.legend = TRUE
)

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/content_analysis_n_count.png"), height = 7, width = 10)
```

# Analyzing Inter-labeler agreement (percentage) and reliability scores

## Percent agreement 

```{r}

# Calculate the percent agreements and then put them with the other information into a dataframe
df_agree <- data.frame(
  "percent_agreement" = c(
    extract_percent_agreement(subset(df, group = "Asian Americans")[, c(3, 6)]),
    extract_percent_agreement(subset(df, group = "Asian Americans")[, c(4, 7)]),
    extract_percent_agreement(subset(df, group != "Asian Americans")[, c(3, 6)]),
    extract_percent_agreement(subset(df, group != "Asian Americans")[, c(4, 7)])
  ),
  "type" = c("Collective gain", "Collective loss", "Collective gain", "Collective loss"),
  "group" = c(rep("Asian Americans", 2), rep("African Americans", 2))
)

# Visualize
df_agree %>%
  mutate(percent_agreement = round(percent_agreement, 0)) %>%
  ggplot(aes(x = group, y = percent_agreement / 100, fill = type)) +
  geom_col(position = "dodge", color = "black") +
  geom_text(aes(label = paste(percent_agreement, "%")), position = position_dodge(width = 0.9), vjust = -0.25) +
  labs(
    title = "Percentage agreement",
    y = "Agreement rate",
    x = "Group",
    fill = "Type"
  ) +
  scale_y_continuous(labels = scales::percent_format(accuracy = 5L)) +
  scale_fill_manual(name = "Type", labels = c("Collective loss", "Collective gain"), values = c("red", "blue"))

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/content_analysis_agreement.png"), height = 7)
```

## Kappa statistics 

### Original
 
```{r extract}

# Extract kappa statistics

df_kappa <- rbind(
  extract_kappa(subset(df, group = "Asian Americans")[, c(3, 6)]),
  extract_kappa(subset(df, group = "Asian Americans")[, c(4, 7)]),
  extract_kappa(subset(df, group != "Asian Americans")[, c(3, 6)]),
  extract_kappa(subset(df, group != "Asian Americans")[, c(4, 7)])
)

df_kappa$type <- c("Collective gain", "Collective loss", "Collective gain", "Collective loss")

df_kappa$group <- c(rep("Asian Americans", 2), rep("African Americans", 2))

# Visualize

df_kappa_plot <- df_kappa %>%
  mutate(kappa = round(kappa, 2)) %>%
  ggplot(aes(x = group, y = kappa, color = type)) +
  geom_pointrange(aes(ymax = upr, ymin = lwr)) +
  geom_text(aes(label = kappa), position = position_dodge(width = 0.9), vjust = -0.25) +
  labs(
    title = "Inter-coder reliability test",
    subtitle = "Including all topics",
    y = "Kappa coefficient",
    x = "Group",
    fill = "Type"
  ) +
  scale_color_manual(name = "Type", labels = c("Collective loss", "Collective gain"), values = c("red", "blue")) +
  coord_flip() +
  ylim(c(0, 1))

dfn <- df %>%
  filter(!Topics_C %in% c("Mismatch", "Others", "Arts", "Recipes", "OTHERS", "Articles in Spanish", "Christian advice"))

dfn_kappa <- rbind(
  extract_kappa(subset(dfn, group = "Asian Americans")[, c(3, 6)]),
  extract_kappa(subset(dfn, group = "Asian Americans")[, c(4, 7)]),
  extract_kappa(subset(dfn, group != "Asian Americans")[, c(3, 6)]),
  extract_kappa(subset(dfn, group != "Asian Americans")[, c(4, 7)])
)

dfn_kappa$type <- c("Collective gain", "Collective loss", "Collective gain", "Collective loss")

dfn_kappa$group <- c(rep("Asian Americans", 2), rep("African Americans", 2))

# Visualize

dfn_kappa_plot <- dfn_kappa %>%
  mutate(kappa = round(kappa, 2)) %>%
  ggplot(aes(x = group, y = kappa, color = type)) +
  geom_pointrange(aes(ymax = upr, ymin = lwr)) +
  geom_text(aes(label = kappa), position = position_dodge(width = 0.9), vjust = -0.25) +
  labs(
    title = "",
    subtitle = "Excluding non-political topics",
    y = "Kappa coefficient",
    x = "Group",
    fill = "Type"
  ) +
  scale_color_manual(name = "Type", labels = c("Collective loss", "Collective gain"), values = c("red", "blue")) +
  coord_flip() +
  ylim(c(0, 1))

df_kappa_plot + dfn_kappa_plot

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/content_analysis_kappa_comp.png"), width = 10, height = 7)
```

## Exporting the summaries   

```{r}

irr_summary <- merge(df_agree, df_kappa)

irr_summary <- irr_summary %>%
  mutate(percent_agreement = percent_agreement / 100)

write.csv(irr_summary, make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/irr_summary.csv"))
```

# Additional exploratory data analyses 

## Construct validity test

```{r}

# Asian Americans
train_asian_plot <- bind_rows(
  mutate(training_asian %>%
    rename(Collective_gain = linked_progress) %>%
    rename(Collective_loss = linked_hurt) %>%
    gather("linked_fate", "value", Collective_gain, Collective_loss) %>%
    filter(value == 1) %>%
    filter(Topics_C != "Mismatch") %>%
    filter(Topics_C != "Others") %>%
    group_by(linked_fate) %>%
    count(Topics_C) %>%
    spread(linked_fate, n) %>%
    mutate(
      Collective_gain = ifelse(is.na(Collective_gain), 0, Collective_gain),
      Collective_loss = ifelse(is.na(Collective_loss), 0, Collective_loss)
    ) %>%
    mutate(difference = Collective_gain - Collective_loss), Measurement = "Maximum"),

  mutate(training_asian %>%
    rename(Collective_gain = linked_progress_gran) %>%
    rename(Collective_loss = linked_hurt_gran) %>%
    gather("linked_fate", "value", Collective_gain, Collective_loss) %>%
    filter(value == 1) %>%
    filter(Topics_C != "Mismatch") %>%
    filter(Topics_C != "Others") %>%
    group_by(linked_fate) %>%
    count(Topics_C) %>%
    spread(linked_fate, n) %>%
    mutate(
      Collective_gain = ifelse(is.na(Collective_gain), 0, Collective_gain),
      Collective_loss = ifelse(is.na(Collective_loss), 0, Collective_loss)
    ) %>%
    mutate(difference = Collective_gain - Collective_loss), Measurement = "Minimum")
)

# African Americans
train_black_plot <- bind_rows(
  mutate(training_black %>%
    rename(Collective_gain = linked_progress) %>%
    rename(Collective_loss = linked_hurt) %>%
    gather("linked_fate", "value", Collective_gain, Collective_loss) %>%
    filter(value == 1) %>%
    filter(Topics_C != "Mismatch") %>%
    filter(Topics_C != "Others") %>%
    group_by(linked_fate) %>%
    count(Topics_C) %>%
    spread(linked_fate, n) %>%
    mutate(
      Collective_gain = ifelse(is.na(Collective_gain), 0, Collective_gain),
      Collective_loss = ifelse(is.na(Collective_loss), 0, Collective_loss)
    ) %>%
    mutate(difference = Collective_gain - Collective_loss), Measurement = "Maximum"),

  mutate(training_black %>%
    rename(Collective_gain = linked_progress_gran) %>%
    rename(Collective_loss = linked_hurt_gran) %>%
    gather("linked_fate", "value", Collective_gain, Collective_loss) %>%
    filter(value == 1) %>%
    filter(Topics_C != "Mismatch") %>%
    filter(Topics_C != "Others") %>%
    group_by(linked_fate) %>%
    count(Topics_C) %>%
    spread(linked_fate, n) %>%
    mutate(
      Collective_gain = ifelse(is.na(Collective_gain), 0, Collective_gain),
      Collective_loss = ifelse(is.na(Collective_loss), 0, Collective_loss)
    ) %>%
    mutate(difference = Collective_gain - Collective_loss), Measurement = "Minimum")
)

# Visualize

ggarrange(
  train_asian_plot %>%
    ggplot(aes(x = fct_reorder(Topics_C, difference), y = difference, fill = Measurement)) +
    geom_col(position = "dodge", color = "black") +
    coord_flip() +
    labs(x = "Topics", y = "# of Collective gain - # of Collective loss", title = "Asian Americans") +
    scale_fill_manual(name = "Measurement", labels = c("Maximum", "Minimum"), values = c("red", "blue")),

  train_black_plot %>%
    ggplot(aes(x = fct_reorder(Topics_C, difference), y = difference, fill = Measurement)) +
    geom_col(position = "dodge", color = "black") +
    coord_flip() +
    labs(x = "Topics", y = "# of Collective gain - # of Collective loss", title = "African Americans") +
    scale_fill_manual(name = "Measurement", labels = c("Maximum", "Minimum"), values = c("red", "blue")),
  common.legend = TRUE
)

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/content_analysis_topics_gran.png"), width = 13, height = 7)
```


# Reading texts 

- Merging the text files with the content analysis 

```{r}

sample_asian <- readRDS(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/raw_data/sample_asian.rds")) %>%
  as.data.frame()

sample_black <- readRDS(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/raw_data/sample_black.rds")) %>%
  as.data.frame()

sample_asian <- bind_cols(sample_asian, training_asian %>% select(linked_progress, linked_hurt, linked_progress_gran, linked_hurt_gran, Topics_C))

sample_black <- bind_cols(sample_black, training_black %>% select(linked_progress, linked_hurt, linked_progress_gran, linked_hurt_gran, Topics_C))
```

- Comparing 100% and 50% data 

```{r}

# Merge

df <- bind_rows(
  mutate(sample_asian, group = "Asian Americans"),
  mutate(sample_black, group = "African Americans")
)

# Select

df <- df %>% select(
  source, year, group, linked_progress, linked_progress_gran,
  linked_hurt, linked_hurt_gran
)

# Mutate

df <- df %>%
  mutate(
    lp_exclusive = ifelse(linked_progress == 1 & linked_hurt == 0, 1, 0),
    lh_exclusive = ifelse(linked_progress == 0 & linked_hurt == 1, 1, 0),
    lf_mixed = ifelse(linked_progress == 1 & linked_hurt == 1, 1, 0)
  ) %>%
  mutate(
    lp_exclusive_gran = ifelse(linked_progress_gran == 1 & linked_hurt_gran == 0, 1, 0),
    lh_exclusive_gran = ifelse(linked_progress_gran == 0 & linked_hurt_gran == 1, 1, 0),
    lf_mixed_gran = ifelse(linked_progress_gran == 1 & linked_hurt_gran == 1, 1, 0)
  )

# Remove numbers and special characters in the source titles

df$source <- gsub("[[:digit:]]", "", df$source)
df$source <- gsub("[[:punct:]]+", "", df$source) %>% trimws()
```

- Non-excluded 

```{r}

# Gather, group by and summary, and mutate

df$group <- factor(df$group, levels = c("Asian Americans", "African Americans"))

df_source <- df %>%
  gather(linked_fate, value, c(lp_exclusive, lh_exclusive, lf_mixed, lp_exclusive_gran, lh_exclusive_gran, lf_mixed_gran)) %>%
  group_by(linked_fate, group) %>%
  summarize_content()

# Create two additional columns for sorting

df_source$type <- ifelse(str_detect(df_source$linked_fate, "gran"), "Minimum", "Maximum")
df_source$linked_fate <- str_replace(df_source$linked_fate, "_gran", "")

# Visualize

plot_ext_cases <- df_source %>%
  filter(linked_fate != "lf_mixed") %>%
  mutate(mean = round(mean, 2)) %>%
  ggplot(aes(x = fct_reorder(type, mean), y = mean, fill = linked_fate)) +
  geom_bar(
    stat = "identity", color = "black",
    position = position_dodge()
  ) +
  geom_errorbar(aes(ymin = lower.ci, ymax = upper.ci),
    width = .2,
    position = position_dodge(.9)
  ) +
  scale_fill_manual(name = "Type", labels = c("Collective loss", "Collective gain"), values = c("red", "blue")) +
  scale_y_continuous(labels = scales::percent_format(accuracy = 5L)) +
  geom_text(aes(label = paste(mean * 100, "%")),
    position = position_dodge(width = 0.9), vjust = -0.25
  ) +
  labs(y = "Proportion", x = "Threshold") +
  facet_wrap(~group)

plot_ext_cases

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/content_data_sources_sub.png"), width = 10, height = 6)
```

# Exporting non-duplicated files 

```{r}

# Non-duplicated
sample_asian <- sample_asian[!duplicated(sample_asian$text), ]
sample_black <- sample_black[!duplicated(sample_black$text), ]

# Check
1008 - nrow(sample_asian)
1008 - nrow(sample_black)

# Export
write.csv(sample_asian, make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/sample_asian.csv"))

write.csv(sample_black, make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/sample_black.csv"))
```
