==> code/03_model_evaluations_comparison.Rmd <==
---
title: "Model evaluations"
author: "Jae Yeon Kim"
html_document:
  toc: True
  theme: united
  number_sections: True
---

## 0. Setup 

```{r}

pacman::p_load(
        tidyverse, # for the tidyverse framework
        ggpubr, # for arranging ggplot2 
        ggthemes, # for fancy ggplot themes
        broom, # for visualizing coefficients
        kableExtra # for nice tables in R markdown
)


# devtools::install_github("jaeyk/makereproducible")
library(makereproducible)

# Import R scripts

script_list <- list.files(paste0(here::here(), "/functions"),
  pattern = "*.r|*.R",
  full.names = TRUE
)


for (i in 1:length(script_list))
{
  source(script_list[[i]])
}

# for publication-friendly theme
theme_set(theme_pubr())
```

# Importing files 

```{r include=FALSE}

merged_model50 <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/merged_model50.csv"))

merged_model100 <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/merged_model100.csv"))

```

# Merging them together 

```{r}

merged_model <- bind_rows(mutate(merged_model50, type = "Minimum"),
                          mutate(merged_model100, type = "Maximum"))

merged_model %>%
  select(-c(content_kappa, content_agreement, X1)) %>%
  filter(resampling == "Upsampling") %>%
  filter(metrices == "Balanced Accuracy") %>%
  filter(models == "Lasso") %>%
  kable() %>%
  kable_styling()
# Summary stat 

merged_model %>% 
  select(models, group, measure, resampling, metrices, rate, type) %>%
  filter(resampling == "None" & type == "Minimum") %>%
  group_by(metrices) %>%
  summarize(min = min(rate),
            max = max(rate),
            mean = mean(rate))

merged_model <- merged_model %>%
  mutate(measure = recode(measure, 
                          "Linked hurt" = "Collective loss",
                          "Linked progress" = "Collective gain"))

```

# Visualizing data 

## Resampling effect  

```{r}

no_resampling_plot100 <- merged_model %>%
  filter(resampling == "None" & type == "Maximum") %>%
  mutate(rate = round(rate, 3)) %>%
  visualize_performance() +
  scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
  labs(title = "ML performances (without upsampling)",
       subtitle = "Maximum",
       x = "Models", 
       y = "Rate") 

no_resampling_plot50 <- merged_model %>%
  filter(resampling == "None" & type == "Minimum") %>%
  mutate(rate = round(rate, 3)) %>%
  visualize_performance() +
  scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
  labs(title = "ML performances (without upsampling)",
       subtitle = "Minimum",
       x = "Models", 
       y = "Rate") 

resampling_plot100 <- merged_model %>%
  filter(resampling != "None" & type == "Maximum") %>%
  mutate(rate = round(rate, 3)) %>%
  visualize_performance() +
  scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
  labs(title = "ML performances (with upsampling)",
       subtitle = "Maximum",
       x = "Models", 
       y = "Rate") 

resampling_plot50 <- merged_model %>%
  filter(resampling != "None" & type == "Minimum") %>%
  mutate(rate = round(rate, 3)) %>%
  visualize_performance() +
  scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
  labs(title = "ML performances (with upsampling)",
       subtitle = "Minimum",
       x = "Models", 
       y = "Rate") 

ggarrange(no_resampling_plot100, resampling_plot100, 
          no_resampling_plot50, resampling_plot50, common.legend = TRUE)

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/ml_performances_comp.png"), width = 14, height = 8)

```

```{r}

no_resampling_plot100 <- merged_model %>%
  filter(resampling == "None" & type == "Maximum") %>%
  mutate(rate = round(rate, 3)) %>%
  visualize_performance() +
  scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
  labs(title = "ML performances (without upsampling)",
       subtitle = "Maximum",
       x = "Models", 
       y = "Rate",
       fill = "Metrics") +
       geom_hline(aes(yintercept = content_agreement), linetype = "dashed") 

no_resampling_plot50 <- merged_model %>%
  filter(resampling == "None" & type == "Minimum") %>%
  mutate(rate = round(rate, 3)) %>%
  visualize_performance() +
  scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
  labs(title = "ML performances (without upsampling)",
       subtitle = "Minimum",
       x = "Models", 
       y = "Rate",
       fill = "Metrics")  +
       geom_hline(aes(yintercept = content_agreement), linetype = "dashed") 

resampling_plot100 <- merged_model %>%
  filter(resampling != "None" & type == "Maximum") %>%
  mutate(rate = round(rate, 3)) %>%
  visualize_performance() +
  scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
  labs(title = "ML performances (with upsampling)",
       subtitle = "Maximum",
       x = "Models", 
       y = "Rate",
       fill = "Metrics") +
       geom_hline(aes(yintercept = content_agreement), linetype = "dashed") 

resampling_plot50 <- merged_model %>%
  filter(resampling != "None" & type == "Minimum") %>%
  mutate(rate = round(rate, 3)) %>%
  visualize_performance() +
  scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
  labs(title = "ML performances (with upsampling)",
       subtitle = "Minimum",
       x = "Models", 
       y = "Rate",
       fill = "Metrics")  +
       geom_hline(aes(yintercept = content_agreement), linetype = "dashed") 

ggarrange(no_resampling_plot100, resampling_plot100, 
          no_resampling_plot50, resampling_plot50, common.legend = TRUE)

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/ml_content_comp.png"), height = 8, width = 14)

```
