==> code/03_model_evaluations_weaker_version.Rmd <==
---
title: "Model evaluations"
author: "Jae Yeon Kim"
html_document:
  toc: True
  theme: united
  number_sections: True
---

# Setup 

```{r}

pacman::p_load(
        tidyverse, # for the tidyverse framework
        ggpubr, # for arranging ggplot2 
        ggthemes, # for fancy ggplot themes
        broom # for visualizing coefficients
)


# devtools::install_github("jaeyk/makereproducible")
library(makereproducible)

# Import R scripts

script_list <- list.files(paste0(here::here(), "/functions"),
  pattern = "*.r|*.R",
  full.names = TRUE
)


for (i in 1:length(script_list))
{
  source(script_list[[i]])
}

# for publication-friendly theme
theme_set(theme_pubr())
```

# Importing files 

```{r include=FALSE}

# From ML models 

# None 

asian_lp_models <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/asian_lp_models_gran.csv"))[,-1]

asian_lh_models <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/asian_lh_models_gran.csv"))[,-1]

black_lp_models <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/black_lp_models_gran.csv"))[,-1]

black_lh_models <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/black_lh_models_gran.csv"))[,-1]

# Resampled 

asian_lp_models_resample <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/asian_lp_models_resample_gran.csv"))[,-1]

asian_lh_models_resample <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/asian_lh_models_resample_gran.csv"))[,-1]

black_lp_models_resample <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/black_lp_models_resample_gran.csv"))[,-1]

black_lh_models_resample <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/black_lh_models_resample_gran.csv"))[,-1]

# From content analysis: Correlation coefficients and Kappa statistics 

irr_summary <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/irr_summary.csv"))[,-1]

```

# Wrangling data

## Putting all ML evals together

## Putting all ML evals together

```{r}

all_evals <- bind_rows(mutate(asian_lp_models, group = "Asian Americans", measure = "Collective gain", resampling = "None"),
          mutate(asian_lh_models, group = "Asian Americans", measure = "Collective loss", resampling = "None"),
          mutate(black_lp_models, group = "African Americans", measure = "Collective gain", resampling = "None"),
          mutate(black_lh_models, group = "African Americans", measure = "Collective loss", resampling = "None"),
          mutate(asian_lp_models_resample, group = "Asian Americans", measure = "Collective gain", resampling = "Upsampling"),
          mutate(asian_lh_models_resample, group = "Asian Americans", measure = "Collective loss", resampling = "Upsampling"),
          mutate(black_lp_models_resample, group = "African Americans", measure = "Collective gain", resampling = "Upsampling"),
          mutate(black_lh_models_resample, group = "African Americans", measure = "Collective loss", resampling = "Upsampling")
          )

# Gather all the models
all_evals <- all_evals %>%
  gather(metrices, rate, c("Accuracy", "Balanced Accuracy"))

```

## Renaming the columns in the content analysis

```{r}

irr_resummary <- irr_summary %>%
  rename(measure = type,
         content_agreement = percent_agreement,
         content_kappa = kappa)

```

## Merging them together

```{r}

# Merging
merged_model <- all_evals %>% right_join(irr_resummary)

# Upper casing col names
names(merged_model) <- tolower(names(merged_model))

write.csv(merged_model, make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/merged_model50.csv"))
```

# Visualizing data 

## Resampling effect  

```{r}

no_resampling_plot <- merged_model %>%
  filter(resampling == "None") %>%
  mutate(rate = round(rate, 3)) %>%
  ggplot(aes(x = models, y = rate, fill = metrices)) +
    geom_col(position = "dodge") +
    scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
    facet_grid(measure ~ group) +
    ylim(c(0, 1)) +
    labs(title = "ML performances (without upsampling)", x = "Models", y = "Rate",
         fill = "Metrics") +
    scale_y_continuous(labels = scales::percent_format(accuracy = 5L)) +
    geom_text(aes(label = paste(rate*100, "%")), position=position_dodge(width=0.9), vjust=-0.25) 

resampling_plot <- merged_model %>%
  filter(resampling != "None") %>%
  mutate(rate = round(rate, 3)) %>%
  ggplot(aes(x = models, y = rate, fill = metrices)) +
    geom_col(position = "dodge") +
    scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
    facet_grid(measure ~ group) +
    ylim(c(0, 1)) +
    labs(title = "ML performances (with upsampling)", x = "Models", y = "Rate",
         fill = "Metrics") +
    scale_y_continuous(labels = scales::percent_format(accuracy = 5L)) +
    geom_text(aes(label = paste(rate*100, "%")), position=position_dodge(width=0.9), vjust=-0.25) 

ggarrange(no_resampling_plot, resampling_plot, ncol = 1, nrow = 2)

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/ml_performances_gran.png"), height = 17, width = 10)

```

```{r}

no_resampling_plot <- merged_model %>%
  filter(resampling == "None") %>%
  mutate(rate = round(rate, 3)) %>%
  ggplot(aes(x = fct_reorder(models, rate), y = rate, fill = metrices)) +
    geom_col(position = "dodge") +
    scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
    geom_hline(aes(yintercept = content_agreement), linetype = "dashed") +
    facet_grid(measure ~ group) +
    ylim(c(0, 1)) +
    labs(title = "ML performances (without upsampling)", x = "Models", y = "Rate",
         fill = "Metrics") +
    scale_y_continuous(labels = scales::percent_format(accuracy = 5L)) +
    geom_text(aes(label = paste(rate*100, "%")), position=position_dodge(width=0.9), vjust=-0.25) 


resampling_plot <- merged_model %>%
  filter(resampling != "None")  %>%
  mutate(rate = round(rate, 3)) %>%
  ggplot(aes(x = fct_reorder(models, rate), y = rate, fill = metrices)) +
    geom_col(position = "dodge") +
    scale_fill_manual(name = "Metrics", labels = c("Accuracy","Balanced accuracy"), values=c("red","blue")) +
    geom_hline(aes(yintercept = content_agreement), linetype = "dashed") +
    facet_grid(measure ~ group) +
    ylim(c(0, 1)) +
    labs(title = "ML performances (with upsampling)", x = "Models", y = "Rate",
         fill = "Metrics") +
    scale_y_continuous(labels = scales::percent_format(accuracy = 5L)) +
    geom_text(aes(label = paste(rate*100, "%")), position=position_dodge(width=0.9), vjust=-0.25) 

ggarrange(no_resampling_plot, resampling_plot, ncol = 1, nrow = 2)

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/ml_content_gran.png"), height = 17, width = 10)
```
