==> code/04_time_series_visualization_weaker_version.Rmd <==
---
title: "Visualizing time series data"
author: "Jae Yeon Kim"
html_document:
  toc: True
  theme: united
  number_sections: True
---

# Setup 

```{r}

pacman::p_load(
        tidyverse, # for the tidyverse framework
        irr, # for calculating inter-coder reliability score
        corrplot, # for visualizing correlation coefficients
        ggpubr, # for arranging ggplots 
        ggthemes, # for fancy ggplot themes
        patchwork, # for arranging ggplots 
        here # reproducibility 
)

# devtools::install_github("jaeyk/makereproducible")
library(makereproducible)

# Import R scripts

script_list <- list.files(paste0(here::here(), "/functions"),
  pattern = "*.r|*.R",
  full.names = TRUE
)


for (i in 1:length(script_list))
{
  source(script_list[[i]])
}

# for publication-friendly theme
theme_set(theme_pubr())
```

# Import files 

```{r include=FALSE}

# Predicted values
asian_lp_data <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/asian_lp_data_gran.csv"))

asian_lh_data <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/asian_lh_data_gran.csv"))

black_lp_data <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/black_lp_data_gran.csv"))

black_lh_data <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/black_lh_data_gran.csv"))

# Labeled articles
asian_sample <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/sample_asian.csv"))

black_sample <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/sample_black.csv"))

# Unlabeled articles
asian_unlabeled <- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/unlabeled_asian.csv"))

black_unlabeled<- read_csv(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/unlabeled_black.csv"))

```

# Wrangle data

## Joinning predicted values and unlabeled articles

```{r}

asian_unlabeled$linked_progress <- asian_lp_data$labeled_linked_progress
asian_unlabeled$linked_hurt <- asian_lh_data$labeled_linked_hurt

black_unlabeled$linked_progress <- black_lp_data$labeled_linked_progress
black_unlabeled$linked_hurt <- black_lh_data$labeled_linked_hurt

```

## Joining labeled and unlabled articles 

```{r}

# Bind them by rows 
asian_full_labeled <- bind_rows(asian_sample, asian_unlabeled)
black_full_labeled <- bind_rows(black_sample, black_unlabeled)

# Drop the first column
asian_full_labeled <- asian_full_labeled[,-1]
black_full_labeled <- black_full_labeled[,-1]

# Remove numbers and special characters in the source titles 
black_full_labeled$source <- gsub('[[:digit:]]', '', black_full_labeled$source) 
black_full_labeled$source <- gsub('[[:punct:]]+', '', black_full_labeled$source) %>% trimws()

```

## Joining Asian American and African American data 

```{r}

# Join the two data 
full_articles <- bind_rows(
  mutate(asian_full_labeled, group = "Asian Americans"),
  mutate(black_full_labeled, group = "African Americans"))

write_csv(full_articles, here("processed_data", "full_articles_gran.csv"))

```

## Check the data types

```{r}

# Check 
glimpse(full_articles)

# Reorder the factor levels 
levels(full_articles$group) <- c("Asian Americans", "African Americans")

```

## Distinguis exclusive from mixed articles 

```{r}

full_articles <- full_articles %>%
  mutate(lp_exclusive = ifelse(linked_progress == 1 & linked_hurt == 0, 1, 0),
         lh_exclusive = ifelse(linked_progress == 0 & linked_hurt == 1, 1, 0),
         lf_mixed = ifelse(linked_progress == 1 & linked_hurt == 1, 1, 0))

names(full_articles)

```

# Visualize data

## Time serieds trends 

```{r}

# Extract year-month
full_articles$year_mon <- format(as.Date(full_articles$date), "%Y-%m")

# Check
full_articles$year_mon[1:10]

```

```{r}

year_trends_asian <- full_articles %>%
  filter(group == "Asian Americans") %>%
  group_by(year) %>%
  visualize_year_trends() +
  labs(subtitle = "Asian Americans")

year_trends_black <- full_articles %>%
  filter(group != "Asian Americans") %>%
  group_by(year) %>%
  visualize_year_trends() +
  labs(subtitle = "African Americans")

month_trends_asian <- full_articles %>%
  filter(group == "Asian Americans") %>%
  group_by(year_mon) %>%
  visualize_month_trends() +
  labs(subtitle = "Asian Americans")

month_trends_black <- full_articles %>%
  filter(group != "Asian Americans") %>%
  group_by(year_mon) %>%
  visualize_month_trends() +
  labs(subtitle = "African Americans")

ggarrange(year_trends_asian, year_trends_black,
          month_trends_asian, month_trends_black, common.legend = TRUE)

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/time_series_plot_gran.png"), height = 10, width = 13)

```

```{r}

time_series_asian <- full_articles %>% filter(group == "Asian Americans") %>% group_by(year_mon) %>% visualize_month_trends() + facet_wrap(~source, ncol = 1) + labs(title = "Monthly trends by group and source", subtitle = "Asian Americans")

time_series_black <- full_articles %>% filter(group == "African Americans") %>% group_by(year_mon) %>% visualize_month_trends() + facet_wrap(~source, ncol = 1) + labs(title = "", subtitle = "African Americans")

ggarrange(time_series_asian, time_series_black, common.legend = TRUE, ncol = 2)

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/time_series_source_gran.png"), height = 13, width = 15)

```

## Matched comparison

```{r}
full_articles %>%
  filter(year >= 1976 & year <= 1981 ) %>%
  group_by(group) %>%
  visualize_matched()

ggsave(make_here("/home/jae/content-analysis-for-evaluating-ML-performances/outputs/matched_comparison_gran.png"), height = 7)

```

# Export the merged file  

```{r}

write.csv(full_articles, make_here("/home/jae/content-analysis-for-evaluating-ML-performances/processed_data/full_articles_gran.csv"))

```
